free study material api

+++ src/main/java/com/aurd/Student/Model/Entity/VideoModel.java
@@ -46,7 +46,7 @@ public class VideoModel {
     private String fee_type;
 
     @Column(name = "course_id",nullable = true)
-    private Integer course_id;
+    private Long course_id;
 
     @Column(name = "sub_subject_id",nullable = true)
     private Integer sub_subject_id;
@@ -95,11 +95,11 @@ public class VideoModel {
         this.inst_id = inst_id;
     }
 
-    public Integer getCourse_id() {
+    public Long getCourse_id() {
         return course_id;
     }
 
-    public void setCourse_id(Integer course_id) {
+    public void setCourse_id(Long course_id) {
         this.course_id = course_id;
     }
